requirements.txt
requests
pytest
pydantic
allure-pytest